requirements.txt
PySide6>=6.6
Pillow>=10.0
pillow-heif>=0.15; platform_system != "Linux"
